Move All Command that only used 1 Subsystem into that Subsystem
