--- src/main/java/frc/robot/commands/ClimberCommands.java
@@ -1,5 +1,0 @@
-package frc.robot.commands;
-
-public class ClimberCommands {
-    // TODO.
-}
