+++ src/main/java/frc/robot/io/swerve_module/SwerveModuleIoTalonFx.java
@@ -39,7 +39,7 @@ import edu.wpi.first.units.Voltage;
  * </p>
  */
 import frc.robot.Constants;
-import frc.robot.subsystems.base.SwerveModule.WheelModuleIndex;
+import frc.robot.subsystems.drive.SwerveModule.WheelModuleIndex;
 
 public class SwerveModuleIoTalonFx implements SwerveModuleIo {
